requirements_FAST.txt
streamlit
nltk
openai
anthropic
python-docx
PyPDF2
pandas
plotly
reportlab
